// frontend/src/app/admin/blogs/models/blog-image-translation.model.ts
export class BlogImageTranslation {
    id?: number;
    image?: string;
    language?: string;
    caption?: string;
    alt_text?: string;
    title_text?: string;
    aria_label?: string;
    long_description?: string;
    
    file_name_override?: string;
    
    updated_at?: string;
    created_at?: string;
}